fix: 사용자 정보 전역 상태 관리 수정
