+++ ZeroPick/vite.config.ts
@@ -2,7 +2,7 @@ import { defineConfig } from 'vite';
 import react from '@vitejs/plugin-react-swc';
 import tsconfigPaths from 'vite-tsconfig-paths';
 import { resolve } from 'path';
-import fs from 'fs';
+/* import fs from 'fs'; */
 
 export default defineConfig({
    plugins: [react(), tsconfigPaths()],
@@ -11,7 +11,7 @@ export default defineConfig({
          '@': resolve(__dirname, 'src'),
       },
    },
-   server: {
+   /*   server: {
       port: 5174,
       proxy: {
          '/api': {
@@ -26,5 +26,5 @@ export default defineConfig({
          key: fs.readFileSync('./cert/localhost+1-key.pem'),
          cert: fs.readFileSync('./cert/localhost+1.pem'),
       },
-   },
+   }, */
 });
